backend/requirements.txt
Flask==2.3.0
Flask-CORS==4.0.0
Flask-SocketIO==5.3.0
pymongo==4.6.0
python-dotenv==1.0.0
requests==2.31.0